Fixed a bug where field names were used as aliases in stead of normal fields
The DB::select() method requires each field as a seperate variable. An array is used for
aliases. So if you pass an array with fields to this methods, it interprets the the first two values as an alias and ignores the rest.

Fixed by using call_user_func_array() method.